Moved A32NX BatTDisplay to fbw-common and added a separate instrument to the A380. Added LVAR for selected displayed battery.
