+++ fbw-a380x/src/base/flybywire-aircraft-a380-842/SimObjects/AirPlanes/FlyByWire_A380_842/panel/panel.cfg
@@ -114,7 +114,7 @@ size_mm=256,128
 pixel_size=256,128
 texture=$BAT
 
-htmlgauge00=A32NX/BAT/bat.html, 0,0,256,128
+htmlgauge00=A380X/BAT/bat.html, 0,0,256,128
 
 [VCockpit15]
 size_mm=1430,1000
