+++ fbw-a380x/src/systems/instruments/src/BAT/config.json
@@ -1,0 +1,10 @@
+{
+  "index": "./index.tsx",
+  "isInteractive": false,
+  "dimensions": {
+    "px": {
+      "width": 256,
+      "height": 128
+    }
+  }
+}
