+++ fbw-a380x/src/systems/instruments/src/BAT/style.scss
@@ -1,0 +1,25 @@
+$colour-main: hsl(40, 100%, 50%);
+
+@font-face {
+  font-family: "Digital";
+  src: url("/Fonts/AirbusBAT.ttf") format("truetype");
+  font-weight: 900;
+  font-style: normal;
+}
+
+.bat-svg {
+  background-color: hsl(55, 82%, 2%);
+  box-shadow: inset 0px 0px 110px -20px rgba(255, 175, 0, 0.25);
+  height: 100%;
+  width: 100%;
+  left: 0;
+  /* TODO: Remove opacity when SAI/ISIS/CLOCK/BAT/ATC/RMP/RUD TRIM is split and not share the same emissive/mesh */
+  opacity: 0.5;
+}
+
+text {
+  fill: $colour-main;
+  font-family: Digital;
+  font-size: 52px;
+  text-anchor: end;
+}
