+++ fbw-a380x/src/systems/instruments/src/BAT/tsconfig.json
@@ -1,0 +1,3 @@
+  {
+    "extends": "../../tsconfig.json"
+  }
