+++ fbw-a380x/src/systems/tsconfig.json
@@ -20,6 +20,7 @@
         "@typings/*": ["../typings"],
         "@flybywiresim/fbw-sdk": ["../../../fbw-common/src/systems/index.ts"],
         "@flybywiresim/clock": ["../../../fbw-common/src/systems/instruments/src/Clock/index.ts"]
+        "@flybywiresim/bat": ["../../../fbw-common/src/systems/instruments/src/BAT/index.ts"]
       }
     }
 }
