+++ fbw-common/src/systems/instruments/src/BAT/index.ts
@@ -1,0 +1,1 @@
+export * from './BatDisplay';
